fs/cromfs and tools/gencromfs:  Various fixes for traversal relative file paths.  Biggest changes is in types used:  Cannot use size_t or mode_t in common structures because they have different sizes on the (64-bit) host and the (32-bit) target.  Use uint32_t instead of size_t for offsets.  Use uint16_t instead of mode_t.

+++ include/nuttx/fs/dirent.h
@@ -108,8 +108,8 @@ struct fs_romfsdir_s
 
 struct fs_cromfsdir_s
 {
-  off_t        cr_firstoffset;         /* Offset to the first entry in the directory */
-  off_t        cr_curroffset;          /* Current offset into the directory contents */
+  uint32_t     cr_firstoffset;         /* Offset to the first entry in the directory */
+  uint32_t     cr_curroffset;          /* Current offset into the directory contents */
 };
 #endif /* CONFIG_FS_CROMFS */
